feat(api): add GET functionality by document collection title

+++ src/main/java/com/dirac/aplicacioningsoftfinal/Model/DatosComputados.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DatosComputados {
-    private int descargasTotales;
+    private long descargasTotales;
     private double valoracionPromedio;
-    private int comentariosTotales;
+    private long comentariosTotales;
 }
